refactor: fix up provider to work with author profile code
