+++ packages/provider/src/author-articles-no-images-base.js
