+++ packages/provider/src/provider.js
@@ -3,7 +3,7 @@ import AuthorProfileProvider, {
   query as authorProfileQuery
 } from "./author-profile";
 import AuthorArticlesNoImagesProvider from "./author-articles-no-images";
-import { query as articleListNoImagesQuery } from "./author-articles-no-images.base";
+import { query as articleListNoImagesQuery } from "./author-articles-no-images-base";
 import AuthorArticlesWithImagesProvider, {
   query as articleListWithImagesQuery
 } from "./author-articles-with-images";
